requirements.txt
minimalmodbus
paho-mqtt
pyserial
PyYAML
pytest
